# ranking: Additional occurrence search optimization