Small optimization improvement. Small negative effect on very small repositories (which are already fast to process), but larger repos show a good boost, so it seems worth it. Stacked on #42931.

| repo                                 | old elapsed | new elapsed | delta             |
| ------------------------------------ | ----------- | ----------- | ----------------- |
| github.com/hashicorp/errwrap         |  19.05ms    |  20.89ms    | 9.65879% increase |
| github.com/hashicorp/go-multierror   |  12.47ms    |  12.10ms    | 67.5221% increase |
| github.com/sourcegraph/lsif-go       |  21.45ms    |  23.56ms    | 9.83683% increase |
| github.com/sourcegraph-testing/etcd  | 160.96ms    | 106.92ms    | 33.5736% decrease |
| github.com/sourcegraph-testing/tidb  | 163.87ms    |  89.54ms    | 45.3591% decrease |
| github.com/sourcegraph-testing/titan |  23.53ms    |  18.79ms    | 20.1445% decrease |
| github.com/sourcegraph-testing/zap   |  28.86ms    |  21.72ms    | 24.7401% decrease |
| github.com/sourcegraph/sourcegraph   | 236.82ms    | 149.12ms    | 37.0323% decrease |
| github.com/vishvananda/netlink       |  27.39ms    |  19.89ms    | 27.3823% decrease |

This change skips the symbol search in the buffer's content if there wouldn't be anything to emit on the output channel. This is the case when all of the paths for a symbol have already been emitted.

## Test plan

Tested locally.